requirements.txt
allure-pytest
allure-python-commons
assertpy
pytest
python-dotenv
requests